snapstate,devicestate: move refreshScheduleManaged into devicestate and hook into snapstate
