+++ overlord/devicestate/export_test.go
@@ -115,4 +115,6 @@ var (
 
 	IncEnsureOperationalAttempts = incEnsureOperationalAttempts
 	EnsureOperationalAttempts    = ensureOperationalAttempts
+
+	RefreshScheduleManaged = refreshScheduleManaged
 )
